app/requirements.txt
Django==3.1.4
gunicorn==20.0.4
psycopg2-binary==2.8.5
djangorestframework==3.12.2
djangorestframework-gis==0.16
django-cors-headers===3.6.0